requirements.txt
Pillow~=10.1.0
fpdf~=1.7.2